Fitting a simple viral infection model with 2 types of drug mechanisms to influenza data

This function fits the simulate_virusandtx_ode model, which is a compartment model using a set of ordinary differential equations. The model describes a simple viral infection system in the presence of drug treatment. The user provides initial conditions and parameter values for the system. The function simulates the ODE using an ODE solver from the deSolve package.

simulate_fit_fludrug(
  U = 1e+05,
  I = 0,
  V = 1,
  dI = 2,
  dV = 4,
  b = 0.001,
  blow = 1e-06,
  bhigh = 1000,
  p = 0.01,
  plow = 1e-06,
  phigh = 1000,
  g = 0.001,
  glow = 0,
  ghigh = 10000,
  e = 0.8,
  fitmodel = 1,
  iter = 100
)

Arguments

U
: initial number of uninfected target cells : numeric

I
: initial number of infected target cells : numeric

V
: initial number of infectious virions : numeric

dI
: rate at which infected cells die : numeric

dV
: rate at which infectious virus is cleared : numeric

b
: rate at which virus infects cells : numeric

blow
: lower bound for infection rate : numeric

bhigh
: upper bound for infection rate : numeric

p
: rate at which infected cells produce virus : numeric

plow
: lower bound for virus production rate : numeric

phigh
: upper bound for virus production rate : numeric

g
: unit conversion factor : numeric

glow
: lower bound for unit conversion factor : numeric

ghigh
: upper bound for unit conversion factor : numeric

e
: drug efficacy (between 0-1) : numeric

fitmodel
: fitting model 1 or 2 : numeric

iter
: max number of steps to be taken by optimizer : numeric

Value

The function returns a list containing the best fit timeseries, the best fit parameters, the data and the AICc for the model.

Details

A simple compartmental ODE models describing an acute viral infection with drug treatment mechanism/model 1 assumes that drug treatment reduces rate of new cell infection mechanism/model 2 assumes that drug treatment reduces rate of new virus production.

Warning

This function does not perform any error checking. So if you try to do something nonsensical (e.g. specify negative parameter or starting values), the code will likely abort with an error message.

See also

Examples

# To run the code with default parameters just call the function:
if (FALSE) result <- simulate_fit_fludrug()
# To apply different settings, provide them to the simulator function, like such:
result <- simulate_fit_fludrug(iter = 5, fitmodel = 1)